requirements.txt
Flask
requests
PyJWT
flask-httpauth
celery
redis
markupsafe
gunicorn